Further work to beautify main and standardize
Added settings for app window size, for saving/restoring between
sessions, and default size. Changed design so that timer data is all
stored in a single file, instead of a separate file for each day. Added
boilerplate code for version update functions. Added update_theme
function. Rejiggered ordering of widgets, sizing, and how they resize so
that it looks nicer when it loads.