Completed tickets 22, 24, 33, started taskfocused
Readjusted timer button intervals so they are a list of 4 elements:
start time, end time, exported, and description.
Changed order of a few functions so that only the version update
functions are above __init__.
Removed the calendar widget, as that will not be used in taskfocused
version.
Added a menubar with options to add a timer or save current data.
Added function to add timer buttons for loaded timer data.
Added function for writing settings/theme/data/passwords to their
respective files.