The Great PEP8 Conversion
- All tabs changed to 4 spaces
- All lines greater than 80 characters chopped down (except in README)
- No actual functional differences

Note that the styling is not very consistent right now, as I am not
used to limiting myself to 80 char lines. That will eventually get
worked out.

Note that this is originally being committed on the partial_work branch
because I don't want to have to deal with merging its differences along
with the tab/space change. So I'll just clobber master when I'm done
with partial_work. Bad form, but all alternatives will cause more
headaches.